refactor(evaluations): move the evaluator display-name map next to the catalog
`evaluatorTempNameMap` lived in a Chakra component, and a legacy route
handler imported it from there to name evaluators in its catalogue
response. That is a server file reaching into the UI tree for a plain
`Record<string, string>`, and it is what stops that route's Hono app
from being wired into the OpenAPI spec task: importing it would pull
Chakra, react-hook-form and the tRPC client into a task bundle that
renders nothing.

The map is a display override on the evaluator catalog, so it belongs
beside the catalog, in the pure module all six consumers already import.

The drawer's integration test mocked the component purely to stub this
map. With the map in a module that costs nothing to import, the mock has
nothing left to do and is gone; the test now runs against the real one.
