+++ platform/app/src/optimization_studio/utils/registryUtils.ts
@@ -1,6 +1,8 @@
-import { evaluatorTempNameMap } from "../../components/checks/EvaluatorSelection";
 import type { useAvailableEvaluators } from "../../hooks/useAvailableEvaluators";
-import type { EvaluatorTypes } from "../../server/evaluations/evaluators";
+import {
+  type EvaluatorTypes,
+  evaluatorTempNameMap,
+} from "../../server/evaluations/evaluators";
 import type { Evaluator, Field } from "../types/dsl";
 
 export { DEFAULT_MAX_TOKENS } from "~/utils/constants";
