+++ platform/app/src/server/routes/evaluations-legacy.ts
@@ -24,7 +24,6 @@ import { nanoid } from "nanoid";
 import { type ZodError, ZodError as ZodErrorClass, z } from "zod";
 import { zodToJsonSchema } from "zod-to-json-schema";
 import { fromZodError } from "zod-validation-error";
-import { evaluatorTempNameMap } from "~/components/checks/EvaluatorSelection";
 import { LEGACY_PAIRWISE_EVALUATOR_TYPE } from "~/experiments-v3/types";
 import { resolveDispatchEvaluatorType } from "~/experiments-v3/utils/normalizeComparison";
 import type { Workflow } from "~/optimization_studio/types/dsl";
@@ -49,6 +48,7 @@ import {
   type EvaluatorDefinition,
   type EvaluatorTypes,
   evaluatorsSchema,
+  evaluatorTempNameMap,
   type SingleEvaluationResult,
 } from "~/server/evaluations/evaluators";
 import { getEvaluatorDefaultSettings } from "~/server/evaluations/getEvaluator";
